+++ iothub/device/src/BaseClient.cs
@@ -7,7 +7,6 @@ namespace Microsoft.Azure.Devices.Client
     using System;
     using System.Collections.Generic;
     using System.Diagnostics;
-    using System.Linq;
 #if NETSTANDARD1_3
     using System.Net.Http;
 #endif
@@ -17,7 +16,6 @@ namespace Microsoft.Azure.Devices.Client
     using Microsoft.Azure.Devices.Client.Extensions;
     using Microsoft.Azure.Devices.Client.Transport;
     using Microsoft.Azure.Devices.Shared;
-    using Newtonsoft.Json.Linq;
     using Microsoft.Azure.Devices.Client.Transport.Mqtt;
     using System.Security.Cryptography.X509Certificates;
 
@@ -504,7 +502,11 @@ TODO: revisit DefaultDelegatingHandler - it seems redundant as long as we have t
             });
         }
 
-        internal Task CompleteInternalAsync(string lockToken)
+        /// <summary>
+        /// Deletes a received message from the device queue
+        /// </summary>
+        /// <returns>The lock identifier for the previously received message</returns>
+        public Task CompleteAsync(string lockToken)
         {
             if (string.IsNullOrEmpty(lockToken))
             {
@@ -515,7 +517,53 @@ TODO: revisit DefaultDelegatingHandler - it seems redundant as long as we have t
             return ApplyTimeout(operationTimeoutCancellationToken => this.InnerHandler.CompleteAsync(lockToken, operationTimeoutCancellationToken.Token));
         }
 
-        internal Task RejectInternalAsync(string lockToken)
+        /// <summary>
+        /// Deletes a received message from the device queue
+        /// </summary>
+        /// <returns>The previously received message</returns>
+        public Task CompleteAsync(Message message)
+        {
+            if (message == null)
+            {
+                throw Fx.Exception.ArgumentNull("message");
+            }
+            // Codes_SRS_DEVICECLIENT_28_015: [The async operation shall retry until time specified in OperationTimeoutInMilliseconds property expire or unrecoverable error(authentication, quota exceed) occurs.]
+            return this.CompleteAsync(message.LockToken);
+        }
+
+        /// <summary>
+        /// Puts a received message back onto the device queue
+        /// </summary>
+        /// <returns>The previously received message</returns>
+        public Task AbandonAsync(string lockToken)
+        {
+            if (string.IsNullOrEmpty(lockToken))
+            {
+                throw Fx.Exception.ArgumentNull("lockToken");
+            }
+            // Codes_SRS_DEVICECLIENT_28_015: [The async operation shall retry until time specified in OperationTimeoutInMilliseconds property expire or unrecoverable error(authentication, quota exceed) occurs.]
+            return ApplyTimeout(operationTimeoutCancellationToken => this.InnerHandler.AbandonAsync(lockToken, operationTimeoutCancellationToken));
+        }
+
+        /// <summary>
+        /// Puts a received message back onto the device queue
+        /// </summary>
+        /// <returns>The lock identifier for the previously received message</returns>
+        public Task AbandonAsync(Message message)
+        {
+            if (message == null)
+            {
+                throw Fx.Exception.ArgumentNull("message");
+            }
+
+            return this.AbandonAsync(message.LockToken);
+        }
+
+        /// <summary>
+        /// Deletes a received message from the device queue and indicates to the server that the message could not be processed.
+        /// </summary>
+        /// <returns>The previously received message</returns>
+        public Task RejectAsync(string lockToken)
         {
             if (string.IsNullOrEmpty(lockToken))
             {
@@ -525,14 +573,18 @@ TODO: revisit DefaultDelegatingHandler - it seems redundant as long as we have t
             return ApplyTimeout(operationTimeoutCancellationToken => this.InnerHandler.RejectAsync(lockToken, operationTimeoutCancellationToken));
         }
 
-        internal Task AbandonInternalAsync(string lockToken)
+        /// <summary>
+        /// Deletes a received message from the device queue and indicates to the server that the message could not be processed.
+        /// </summary>
+        /// <returns>The lock identifier for the previously received message</returns>
+        public Task RejectAsync(Message message)
         {
-            if (string.IsNullOrEmpty(lockToken))
+            if (message == null)
             {
-                throw Fx.Exception.ArgumentNull("lockToken");
+                throw Fx.Exception.ArgumentNull("message");
             }
-            // Codes_SRS_DEVICECLIENT_28_015: [The async operation shall retry until time specified in OperationTimeoutInMilliseconds property expire or unrecoverable error(authentication, quota exceed) occurs.]
-            return ApplyTimeout(operationTimeoutCancellationToken => this.InnerHandler.AbandonAsync(lockToken, operationTimeoutCancellationToken));
+            // Codes_SRS_DEVICECLIENT_28_017: [The async operation shall retry until time specified in OperationTimeoutInMilliseconds property expire or unrecoverable error(authentication, quota exceed) occurs.]
+            return this.RejectAsync(message.LockToken);
         }
 
         /// <summary>
